\name{FullSample}
\alias{FullSample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Function implementing Gibbs Sampler with old-version interface
}
\description{
	This function implements Gibbs Sampler for estimating model 
	parameters, but with the old version interface. This function 
	remains callable just for backward compatibility, and not going
	to be used by new users.
}

\details{
	This function not going to be used by new users.
}

\author{
	Linghao SHEN <sl013@ie.cuhk.edu.hk>
}
